Stable-like without inventory & sell module

+++ productswindow.cpp
@@ -1,9 +1,5 @@
 #include "productswindow.h"
 #include "ui_productswindow.h"
-#include "sql.h"
-#include <QtSql>
-#include <QtDebug>
-#include <QObject>
 
 productsWindow::productsWindow(QWidget *parent) :
     QDialog(parent),
@@ -11,6 +7,7 @@ productsWindow::productsWindow(QWidget *parent) :
 {   
     ui->setupUi(this);
     ui->tableView->horizontalHeader()->setStretchLastSection(true);
+    ui->tableView->horizontalHeader()->setSectionResizeMode(QHeaderView::ResizeToContents);
     updateProductList();
 }
 
@@ -19,7 +16,7 @@ productsWindow::~productsWindow()
     delete ui;
 }
 
- void productsWindow::updateProductList(){
+void productsWindow::updateProductList(){
     sql conn;
     QSqlQueryModel *modal = new QSqlQueryModel;
     conn.dbOpen();
@@ -46,7 +43,7 @@ void productsWindow::on_button_addNew_clicked()
 {
     QString name = ui->input_name->text();
     QString unit = ui->input_unit->text();
-    float price = ui->input_price->text().toFloat();
+    double price = ui->input_price->text().toDouble();
     int tax = ui->input_tax->text().toInt();
 
     //qDebug() << "Name: " << name << " Unit: " << unit << " Price: " << price << " Tax: " << tax; //ok
@@ -90,24 +87,22 @@ void productsWindow::on_button_removeSelected_clicked()
     if(ui->combo_select->currentText()=="ID"){
         qDebug() << "Nie wybrano rekordu!";
     }else{
-        int id = ui->combo_select->currentText().toInt();
-        //qDebug() << id;
-        sql conn;
-        conn.dbOpen();
-        QSqlQuery *query = new QSqlQuery(conn.db);
-
-        query->prepare("delete from produkty where id=:id");
-        query->bindValue(":id",id);
-        query->exec();
-        conn.dbClose();
-
-        updateProductList();
+        QMessageBox::StandardButton reply;
+        reply = QMessageBox::question(this, "Usuwanie rekordu", "Czy napewno chcesz usunąć ten rekord?", QMessageBox::Yes|QMessageBox::No);
+        if (reply == QMessageBox::Yes) {
+            int id = ui->combo_select->currentText().toInt();
+            sql conn;
+            conn.dbOpen();
+            QSqlQuery *query = new QSqlQuery(conn.db);
+            query->prepare("delete from produkty where id=:id");
+            query->bindValue(":id",id);
+            query->exec();
+            conn.dbClose();
+            updateProductList();
+        } else {
+            updateProductList();
+        }
     }
-}
-
-void productsWindow::on_pushButton_clicked()
-{
-    productsWindow::close();
 }
 
 void productsWindow::on_tableView_clicked()
@@ -119,4 +114,17 @@ void productsWindow::on_tableView_clicked()
     if ( comboIndex != -1 ) { // -1 for not found
        ui->combo_select->setCurrentIndex(comboIndex);
     }
+}
+
+void productsWindow::on_button_return_clicked()
+{
+    productsWindow::close();
+}
+
+void productsWindow::on_button_newProduct_clicked()
+{
+    addProducts addProducts;
+    addProducts.setModal(true);
+    addProducts.exec();
+    updateProductList();
 }
